Finished probability class and nearly finished updated standard inf class
Standard inf class can be run for a single batch in entirety with multiprocessing, but one cannot currently combine samples from separate runs. Additionally, there is a bug where sephamore objects are being leaked from the multiprocessing, but it is not impacting the ability to run the class or the results therein.