Improvements to DM param to sigmav converter method and standard inf combine module
For the sigmav converter a new method was added to the base DM spectra class that allows efficient filtering of the unique hyperparameters, knowing that you have a mesh including all the energy values. i.e. Specifically when you know you need to integrate over spectra for a mesh of hyperparameter values that may or may not be a mesh.

The combing script is just that the release of the hyperparameter data for a specific batch of events was outside the try and except statement. Thus if the try and except behaviour was actually used, then it would fail anyway :(.