Basic stick-breaking working as well as it could be
The basics of the mixture modelling with the dark matter annihilation ratios converted into stick-breaking dirichlet process fractions is working as well as it can. Turns out the stick-breaking process exerts a heck of an induced prior on the dirichlet weights/actual annihilation ratios that could be biasing current results and would result in unreliable annihilation ratio outputs.

I have begun implementing a 'tree' structure that will allow more complex mixture modelling to be used natively and also to start working on using standard dirichlet priors/structures in modelling to allow more natural priors to be used.